Added more helpful output for server and client for debugging.
Packets sent and received are displayed on-screen for both client and
server as packets per second. Server now correctly transmits every 2
ticks. Currently working on defining PacketTypes as an enum.
PROBLEM, client-side has some firewall issues with udp when connecting
to server. My laptop can connect properly while firewall is active.

+++ Boop Server/src/Math/Bitmaths.java
@@ -49,11 +49,11 @@ public final class Bitmaths {
 		return ByteBuffer.wrap(bytes).asFloatBuffer().array();
 	}
 	
-	public static float bytesToInt(byte[] bytes) {
+	public static int bytesToInt(byte[] bytes) {
 		return ByteBuffer.wrap(bytes).getInt();
 	}
 	
-	public static float bytesToInt(byte[] bytes, int offset) {
+	public static int bytesToInt(byte[] bytes, int offset) {
 		if (offset > bytes.length - 4) return -1;
 		return ByteBuffer.wrap(bytes).getInt(offset);
 	}
